move dog files into other folder to prep for redux
