+++ src/App/Features/Dogs/DogCard.tsx
@@ -1,9 +1,9 @@
-import type { Dog } from '../types/types';
-import profileImg from '../assets/profile-default.png';
+import type { Dog } from '../../../types/types';
+import profileImg from '../../../assets/profile-default.png';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faX } from '@fortawesome/free-solid-svg-icons';
 import { Link } from 'react-router-dom';
-import { calculateAge } from '../utils/age-utils';
+import { calculateAge } from '../../../utils/age-utils';
 
 interface DogCardProps {
   dog: Dog;
