+++ src/App/Features/Dogs/DogDetail.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from 'react';
 import { generateClient } from 'aws-amplify/data';
-import type { Schema } from '../../amplify/data/resource';
-import type { Dog } from '../types/types';
+import type { Schema } from '../../../../amplify/data/resource';
+import type { Dog } from '../../../types/types';
 import { useParams } from 'react-router-dom';
-import { calculateAge } from '../utils/age-utils';
-import profileImg from '../assets/profile-default.png';
+import { calculateAge } from '../../../utils/age-utils';
+import profileImg from '../../../assets/profile-default.png';
 
 function DogDetail() {
   const client = generateClient<Schema>();
