+++ src/main.tsx
@@ -2,15 +2,15 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import App from './App.tsx';
-import DogList from './components/DogList.tsx';
-import DogDetail from './components/DogDetail.tsx';
+import DogList from './App/Features/Dogs/DogList.tsx';
+import DogDetail from './App/Features/Dogs/DogDetail.tsx';
 import ErrorPage from './components/ErrorPage.tsx';
 import './index.css';
 import { Authenticator } from '@aws-amplify/ui-react';
 import { Amplify } from 'aws-amplify';
 import outputs from '../amplify_outputs.json';
 import '@aws-amplify/ui-react/styles.css';
-import FormCreateDog from './components/FormCreateDog.tsx';
+import FormCreateDog from './App/Features/Dogs/FormCreateDog.tsx';
 
 Amplify.configure(outputs);
